refactor: rework adapter
