+++ packages/mp-loader/src/extensionAdapter.js
@@ -1,15 +1,17 @@
-module.exports = function(mpType) {
-  if (mpType === 'weixin') {
-    return {
-      CSS_EXT: '.wxss',
-      TEMPLATE_EXT: '.wxml',
-      CONFIG_EXT: 'json'
-    };
-  } else {
-    return {
-      CSS_EXT: '.acss',
-      TEMPLATE_EXT: '.axml',
-      CONFIG_EXT: 'json'
-    };
+module.exports = function(type) {
+  switch (type) {
+    case 'weixin':
+      return {
+        CSS_EXT: '.wxss',
+        TEMPLATE_EXT: '.wxml',
+        CONFIG_EXT: 'json'
+      };
+    case 'alipay':
+    default:
+      return {
+        CSS_EXT: '.acss',
+        TEMPLATE_EXT: '.axml',
+        CONFIG_EXT: 'json'
+      };
   }
 };
