+++ packages/mp-loader/src/transpiler/createGenerator.js
