--- packages/mp-loader/src/transpiler/transformModules/alipay/bind.js
@@ -1,7 +1,0 @@
-const { createTransformNode } = require('../../transformCreator/bind');
-
-const IS_DETECTIVE = /^a\:/;
-
-module.exports = {
-  transformNode: createTransformNode(IS_DETECTIVE)
-};
