--- packages/mp-loader/src/transpiler/transformModules/alipay/events.js
@@ -1,7 +1,0 @@
-const { createTransformNode } = require('../../transformCreator/events');
-
-const LISTENER_ACTION = 'on';
-
-module.exports = {
-  transformNode: createTransformNode(LISTENER_ACTION)
-};
