--- packages/mp-loader/src/transpiler/transformModules/alipay/web-view.js
@@ -1,7 +1,0 @@
-const { createTransformNode } = require('../../transformCreator/web-view');
-
-const MESSAGE_ATTRIBUTE = 'onMessage';
-
-module.exports = {
-  transformNode: createTransformNode(MESSAGE_ATTRIBUTE)
-};
