--- packages/mp-loader/src/transpiler/transformModules/weapp/bind.js
@@ -1,7 +1,0 @@
-const { createTransformNode } = require('../../transformCreator/bind');
-
-const IS_DETECTIVE = /^wx\:/;
-
-module.exports = {
-  transformNode: createTransformNode(IS_DETECTIVE)
-};
