--- packages/mp-loader/src/transpiler/transformModules/weapp/key.js
@@ -1,7 +1,0 @@
-const { createTransformNode } = require('../../transformCreator/key');
-
-const ATTR_KEY = 'wx:key';
-
-module.exports = {
-  transformNode: createTransformNode(ATTR_KEY)
-};
